requirements.txt
test-common==1.3.0.post3
python-gnupg==0.5.0